chore: refactor registrar service and update docs
