+++ schemas/CreateRegistrarConfigDto.schema.json
@@ -36,7 +36,7 @@
     },
     "registrationCertificateDefaults": {
       "nullable": true,
-      "description": "Optional default values merged into registration certificate creation requests (for example privacy_policy, support_uri, provided_attestations)",
+      "description": "Optional default values merged into registration certificate creation requests (for example privacy_policy, support_uri)",
       "additionalProperties": true,
       "type": "object",
       "allOf": [
