+++ schemas/RegistrationCertificateBody.schema.json
@@ -32,9 +32,5 @@
       }
     }
   },
-  "required": [
-    "privacy_policy",
-    "support_uri"
-  ],
   "additionalProperties": false
 }
